+++ lldb/test/API/tools/lldb-dap/attach-commands/Makefile
@@ -1,0 +1,3 @@
+C_SOURCES := main.c
+
+include Makefile.rules
