+++ lldb/test/API/tools/lldb-dap/attach-commands/main.c
@@ -2,6 +2,7 @@
 #include <stdio.h>
 #ifdef _WIN32
 #include <process.h>
+#include <windows.h>
 #else
 #include <unistd.h>
 #endif
@@ -19,9 +20,11 @@ int main(int argc, char const *argv[]) {
     fclose(f);
   }
 
-  // Wait on input from stdin.
-  getchar();
-
   printf("pid = %i\n", getpid());
-  return 0;
+#ifdef _WIN32
+  Sleep(10 * 1000);
+#else
+  sleep(10);
+#endif
+  return 0; // breakpoint 1
 }
